+++ .eslintrc.js
@@ -1,5 +1,3 @@
-const { peerDependencies } = require('./package.json');
-
 module.exports = {
   parser: '@typescript-eslint/parser',
   parserOptions: {
@@ -9,10 +7,11 @@ module.exports = {
   plugins: ['@typescript-eslint', 'react-hooks'],
   extends: [
     'airbnb',
+    'airbnb/hooks',
     'plugin:@typescript-eslint/recommended',
     'plugin:@typescript-eslint/recommended-requiring-type-checking',
+    'plugin:react/recommended',
     'prettier',
-    'prettier/@typescript-eslint',
   ],
   settings: {
     'import/resolver': {
@@ -25,34 +24,25 @@ module.exports = {
     browser: true,
   },
   rules: {
-    '@typescript-eslint/array-type': 'off',
-    '@typescript-eslint/explicit-member-accessibility': ['warn', { accessibility: 'no-public' }],
-    '@typescript-eslint/no-empty-function': 'error',
-    '@typescript-eslint/no-empty-interface': 'off',
-    '@typescript-eslint/no-explicit-any': 'off',
-    '@typescript-eslint/no-parameter-properties': 'off',
-    '@typescript-eslint/no-use-before-define': 'off',
-    '@typescript-eslint/no-useless-constructor': 'error',
-    '@typescript-eslint/unbound-method': 'off',
+    'arrow-body-style': 'off',
     'import/extensions': 'off',
-    'import/no-unresolved': ['error', { ignore: Object.keys(peerDependencies) }],
+    'import/no-unresolved': 'off',
     'import/prefer-default-export': 'off',
-    'jsx-a11y/accessible-emoji': 'off',
     'jsx-a11y/anchor-is-valid': 'off',
-    'jsx-a11y/label-has-associated-control': 'off',
-    'no-continue': 'off',
-    'no-empty-function': 'off',
     'no-plusplus': 'off',
-    'no-useless-constructor': 'off',
+    'no-restricted-syntax': 'off',
+    'no-shadow': 'off',
+    'no-use-before-define': 'off',
+    'no-void': 'off',
     'prefer-arrow-callback': 'off',
-    'react-hooks/exhaustive-deps': 'warn',
-    'react-hooks/rules-of-hooks': 'error',
-    'react/jsx-curly-brace-presence': 'off',
+    'react/display-name': 'off',
+    'react/forbid-prop-types': 'off',
     'react/jsx-filename-extension': ['warn', { extensions: ['.tsx'] }],
     'react/jsx-one-expression-per-line': 'off',
     'react/jsx-props-no-spreading': 'off',
-    'react/jsx-wrap-multilines': 'off',
+    'react/no-unescaped-entities': 'off',
     'react/prop-types': 'off',
+    'react/react-in-jsx-scope': 'off',
     'react/require-default-props': 'off',
   },
 };