+++ src/Slider/Slider.tsx
@@ -1,13 +1,13 @@
-import * as React from 'react';
+import React from 'react';
 
-import { OpaqueInterpolation, animated } from 'react-spring';
+import { MotionStyle, MotionValue, motion, useMotionTemplate, useTransform } from 'framer-motion';
 
 import { ItemWrapper } from './ItemWrapper';
 
 type Props = {
   children?: React.ReactNode;
   count: number;
-  index: OpaqueInterpolation<number>;
+  index: MotionValue<number>;
   margin: number;
   style?: React.CSSProperties;
 } & React.DetailedHTMLProps<React.HTMLAttributes<HTMLDivElement>, HTMLDivElement>;
@@ -20,7 +20,7 @@ export const Slider = React.forwardRef<HTMLDivElement, Props>(function Slider(
   const tail = childrenCount - count;
   const frameWidth = 100 / childrenCount;
 
-  const startIndex = index.interpolate((value) => {
+  const startIndex = useTransform(index, (value) => {
     if (!tail) {
       return 0;
     }
@@ -34,7 +34,7 @@ export const Slider = React.forwardRef<HTMLDivElement, Props>(function Slider(
     );
   });
 
-  const translate = index.interpolate((value) => {
+  const translate = useTransform(index, (value) => {
     if (!tail) {
       return 0;
     }
@@ -51,9 +51,11 @@ export const Slider = React.forwardRef<HTMLDivElement, Props>(function Slider(
     overflow: 'hidden',
   };
 
-  const sliderStyle: React.CSSProperties = {
+  const transform = useMotionTemplate`translateX(-${translate}%)`;
+
+  const sliderStyle: MotionStyle = {
     height: '100%',
-    transform: translate.interpolate((value) => `translateX(${-value}%)`),
+    transform,
     willChange: 'transform',
     display: 'flex',
     width: `calc(((100% - ${(count - 1) * margin}px) / ${count} + ${margin}px)*${childrenCount})`,
@@ -61,7 +63,7 @@ export const Slider = React.forwardRef<HTMLDivElement, Props>(function Slider(
 
   return (
     <div ref={ref} style={containerStyle} {...props}>
-      <animated.div style={sliderStyle}>
+      <motion.div style={sliderStyle}>
         {React.Children.map(children, (child, i) => (
           <ItemWrapper
             childrenCount={childrenCount}
@@ -72,7 +74,7 @@ export const Slider = React.forwardRef<HTMLDivElement, Props>(function Slider(
             {child}
           </ItemWrapper>
         ))}
-      </animated.div>
+      </motion.div>
     </div>
   );
 });